[app:index] Move `client/app.tsx` to `app/index.tsx`
[server:index] Update import from `client/app` to `app/index`
[.eslintrc.js] Update alias mappings: remove mapping for `client` and
add mapping for `app`
[jest.config.json] Update moduleNameMapper config: remove mapping for
`client` and add mapping for `app`
[tsconfig.json`] Update paths mapping config: remove mapping for
`client` and add mapping for `app`
[webpack.config.js] Update alias mapping config: remove mapping for
`client` and add mapping for `app`
[webpack.config.js] Add app dir to Nodemon plugin watch list
