+++ client/entry.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { hydrateRoot } from 'react-dom/client';
 
-import App from './app';
+import App from 'app/index';
 
 hydrateRoot(document, <App />);
